# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI 智能下单与生产协同平台 API
  version: 1.0.0
  description: |
    牙科定制工厂三端系统（医生端 / 客服端 / 生产端 / 管理端）后端 API 规范。

    ## 认证方式
    所有接口（除登录外）必须携带 Authorization: Bearer {access_token}。
    服务端对每个接口强制校验角色权限，禁止仅靠前端隐藏。

    ## 角色枚举
    - DOCTOR：医生
    - CS：客服
    - WORKER：技工/生产人员
    - ADMIN：超级管理员

    ## 通用响应格式
    所有接口统一返回：
    { "code": 200, "msg": "success", "data": {} }
    错误时：
    { "code": 4xx/5xx, "msg": "错误说明", "data": null }
  license:
    name: Proprietary internal API
    url: https://yourdomain.com/legal/proprietary-api-license
servers:
- url: https://api.yourdomain.com/v1
  description: 正式环境
- url: https://test-api.yourdomain.com/v1
  description: 测试环境
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CommonResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        msg:
          type: string
          example: success
        data: {}
...
    CheckRecordRequest:
      type: object
      required:
      - node_instance_id
      - check_type
      - is_pass
      properties:
        node_instance_id:
          type: integer
        check_type:
          type: integer
          enum:
          - 1
          - 2
          description: 1=入检, 2=出检
        is_pass:
          type: boolean
        remark:
          type: string
        rework_to_node_id:
          type: integer
          description: 出检不通过时必填
    CheckRecordResponse:
      type: object
      properties:
        check_id:
          type: integer
        node_instance_id:
          type: integer
        check_type:
          type: integer
          enum:
          - 1
          - 2
        result:
          type: string
          enum:
          - PASS
          - FAIL
        rework_id:
          type: integer
          nullable: true
...
  responses:
...
    Unauthorized:
      description: 未登录或 Token 无效
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    Forbidden:
      description: 无权限或数据范围拒绝
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    Conflict:
      description: 业务状态冲突或幂等冲突
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
...
security:
- BearerAuth: []
paths:
...
  "/check-records":
    post:
      tags:
      - CheckRecord
      summary: 提交入检或出检记录
      description: |
        权限：WORKER / ADMIN
        入检（check_type=1）：当前节点状态必须为 READY。
        出检（check_type=2）：当前节点状态必须为 COMPLETED。
        出检不通过时 rework_to_node_id 必填，系统将指定节点状态回退为 READY。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CheckRecordRequest"
      responses:
        '200':
          description: 记录提交成功
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/CheckRecordResponse"
        '400':
          description: 顺序校验失败（如未入检就出检）
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: postCheckRecords
  "/check-records/{nodeInstanceId}":
    get:
      tags:
      - CheckRecord
      summary: 获取某工序节点的入检/出检记录列表
      description: 权限：WORKER / CS / ADMIN
      parameters:
      - name: nodeInstanceId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 入检/出检记录列表
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        "$ref": "#/components/schemas/CheckRecordResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: getCheckRecordsByNodeInstanceId
  "/reworks":
    get:
      tags:
      - CheckRecord
      summary: 查询返工记录列表
      description: |
        权限：WORKER / CS / ADMIN。
        WORKER 仅返回来源节点或目标节点分配给本人的返工记录；CS / ADMIN 可按状态或订单筛选内部返工记录。
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - PENDING
          - IN_PROGRESS
          - DONE
      - name: order_id
        in: query
        required: false
        schema:
          type: integer
      - name: has_impacted_nodes
        in: query
        required: false
        description: 为 true 时仅返回影响过后续节点的返工记录；为 false 时仅返回未影响后续节点的返工记录。
        schema:
          type: boolean
      responses:
        '200':
          description: 返工记录列表
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        "$ref": "#/components/schemas/ReworkRecordResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: getReworks
...
  "/production/quality/summary":
    get:
      tags:
      - CheckRecord
      summary: 获取生产端质量与返工汇总
      description: |
        权限：WORKER / CS / ADMIN。
        用于生产端质量模块展示总返工率、内返率、外返率、一次通过率、终检通过率、投诉率和退货率。
        医生端不能读取该内部生产质量汇总；投诉率和退货率当前因缺少事实表返回 0。
      parameters:
      - name: product_type
        in: query
        required: false
        description: 可选产品类型筛选；前端默认不传，测试可用它隔离历史数据。
        schema:
          type: string
      responses:
        '200':
          description: 生产质量汇总
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/ProductionQualitySummary"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: getProductionQualitySummary
  "/production/equipment/summary":
    get:
      tags:
      - CheckRecord
      summary: 获取生产端设备管理汇总
      description: |
        权限：WORKER / CS / ADMIN。
        用于生产端设备管理页展示设备台账、设备状态、保养计划、故障报修、停机时长和设备稼动率。
        医生端不能读取该内部生产设备汇总。
      parameters:
      - name: equipment_code_prefix
        in: query
        required: false
        description: 可选设备编号前缀筛选；前端默认不传，测试可用它隔离历史数据。
        schema:
          type: string
      responses:
        '200':
          description: 生产设备汇总
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/ProductionEquipmentSummary"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: getProductionEquipmentSummary
  "/production/equipment":
    post:
      tags:
      - CheckRecord
      summary: 新增生产设备台账
      description: |
        权限：WORKER / ADMIN。
        一期用于生产端人工维护设备台账；不接真实 IoT，不代表自动采集设备状态。
        医生端不能写入内部生产设备数据。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ProductionEquipmentRequest"
      responses:
        '200':
          description: 已创建的设备台账
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/ProductionEquipment"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: createProductionEquipment
  "/production/equipment/{equipmentCode}/events":
    post:
      tags:
      - CheckRecord
      summary: 登记生产设备事件
      description: |
        权限：WORKER / ADMIN。
        一期用于人工登记保养计划、故障报修或停机事件；不做复杂审批或 IoT 联动。
        医生端不能写入内部生产设备事件。
      parameters:
      - name: equipmentCode
        in: path
        required: true
        description: 设备编号。
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ProductionEquipmentEventRequest"
      responses:
        '200':
          description: 已登记的设备事件
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/ProductionEquipmentEvent"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: createProductionEquipmentEvent
  "/production/material-exceptions/summary":
    get:
      tags:
      - CheckRecord
      summary: 获取生产端物料异常汇总
      description: |
        权限：WORKER / CS / ADMIN。
        用于生产端物料异常页展示缺料、错料、批次异常、材料损耗、处理状态和责任归属。
        医生端不能读取该内部生产物料异常汇总。
      parameters:
      - name: exception_no_prefix
        in: query
        required: false
        description: 可选物料异常编号前缀筛选；前端默认不传，测试可用它隔离历史数据。
        schema:
          type: string
      responses:
        '200':
          description: 生产物料异常汇总
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/ProductionMaterialExceptionSummary"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: getProductionMaterialExceptionSummary
  "/production/material-exceptions":
    post:
      tags:
      - CheckRecord
      summary: 登记生产物料异常
      description: |
        权限：WORKER / ADMIN。
        一期用于生产端人工登记缺料、错料、批次异常或材料损耗；不做库存扣减、采购补料、供应商协同或 WMS 联动。
        医生端不能写入内部生产物料异常。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ProductionMaterialExceptionRequest"
      responses:
        '200':
          description: 已登记的物料异常
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/ProductionMaterialException"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: createProductionMaterialException
  "/production/material-exceptions/{exceptionNo}/status":
    put:
      tags:
      - CheckRecord
      summary: 更新生产物料异常处理状态
      description: |
        权限：WORKER / ADMIN。
        一期只做人工更新 PENDING / IN_PROGRESS / CLOSED 处理状态；不触发库存、采购、供应商或财务系统。
        医生端不能更新内部生产物料异常。
      parameters:
      - name: exceptionNo
        in: path
        required: true
        description: 物料异常编号。
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ProductionMaterialExceptionStatusRequest"
      responses:
        '200':
          description: 已更新的物料异常
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/ProductionMaterialException"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: updateProductionMaterialExceptionStatus
  "/production/safety-environment/summary":
    get:
      tags:
      - CheckRecord
      summary: 获取生产端安环管理汇总
      description: |
        权限：WORKER / CS / ADMIN。
        用于生产端安环管理页展示安全巡检、隐患整改、环境记录、PPE/设备安全提醒和安环事件统计。
        医生端不能读取该内部生产安环汇总。
      parameters:
      - name: event_no_prefix
        in: query
        required: false
        description: 可选安环事件编号前缀筛选；前端默认不传，测试可用它隔离历史数据。
        schema:
          type: string
      responses:
        '200':
          description: 生产安环汇总
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/ProductionSafetyEnvironmentSummary"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: getProductionSafetyEnvironmentSummary
  "/production/safety-environment/events":
    post:
      tags:
      - CheckRecord
      summary: 登记生产端安环事件
      description: |
        权限：WORKER / ADMIN。
        用于生产端安环管理页人工登记安全巡检、隐患整改、环境记录或 PPE/设备安全提醒。
        医生端不能写入内部生产安环事件；本接口不接真实环境采集硬件、不做完整安环审批流。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ProductionSafetyEnvironmentEventRequest"
      responses:
        '200':
          description: 已登记的安环事件
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/ProductionSafetyEnvironmentEvent"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: createProductionSafetyEnvironmentEvent
...